Move test classes for classes Diff and ChangedFile to the model.git test package
